Update vector f64 implementation (c)

+++ _cpp/_vector/_vector_float64.cpp
@@ -1,26 +1,26 @@
 #include <vector>
 
 extern "C" {
-    std::vector<int64_t>* _create(){
-        return new std::vector<int64_t>;
+    std::vector<double>* _create(){
+        return new std::vector<double>;
     }
-    std::vector<int64_t>* _from_arr(int64_t* p, int size) {
-        std::vector<int64_t>* v = new std::vector<int64_t>;
+    std::vector<double>* _from_arr(double* p, int size) {
+        std::vector<double>* v = new std::vector<double>;
         for (int i = 0; i < size; i++) {
             v->push_back(p[i]);
         }
         return v;
     }
-    void _delete(std::vector<int64_t>* v){
+    void _delete(std::vector<double>* v){
         delete v;
     }
-    int64_t _size(std::vector<int64_t>* v){
+    double _size(std::vector<double>* v){
         return v->size();
     }
-    int64_t _get(std::vector<int64_t>* v, int i){
+    double _get(std::vector<double>* v, int i){
         return v->at(i);
     }
-    void _push_back(std::vector<int64_t>* v, int64_t i){
+    void _push_back(std::vector<double>* v, double i){
         v->push_back(i);
     }
 }
